增加适配
通过修改导航栏配置，以及一些页面的视图坐标位置，达到适配ios6，ios7的目的

+++ Calendar/JBSXRCUnitTileView.m
@@ -82,7 +82,7 @@
     if (eventsCount == 0) {
         self.eventCountLabel.text = @"";
     } else {
-        self.eventCountLabel.text = [NSString stringWithFormat:@"%i", eventsCount];
+        self.eventCountLabel.text = [NSString stringWithFormat:@"%li", (long)eventsCount];
     }
 }
